requirements.txt
fastapi
uvicorn
python-multipart
pillow
joblib
scikit-learn
gunicorn
jinja2
